src/main/java/com/fredsonchaves07/moviecatchapi/domain/useCases/CreateUserUseCase.java
package com.fredsonchaves07.moviecatchapi.domain.useCases;

import com.fredsonchaves07.moviecatchapi.domain.dto.CreateUserDTO;
import com.fredsonchaves07.moviecatchapi.domain.dto.UserDTO;
import com.fredsonchaves07.moviecatchapi.domain.entities.User;
import com.fredsonchaves07.moviecatchapi.domain.repositories.UserRepository;
import com.fredsonchaves07.moviecatchapi.domain.useCases.exceptions.EmailOrPasswordInvalid;

public class CreateUserUseCase {

    private UserRepository userRepository;

    public CreateUserUseCase(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public UserDTO execute(CreateUserDTO createUserDTO) {
        String name = createUserDTO.getName();
        String email = createUserDTO.getEmail();
        String password = createUserDTO.getPassword();
        User user = new User(name, email, password);
        return new UserDTO(user);
    }

    private boolean isEmailAndPasswordValid(String email, String password) {
        if (!emailIsValid(String email)) {
            throw new EmailOrPasswordInvalid("Email or password invalid");
        }
    }

    private boolean emailIsValid(String email) {
        return false;
    }
}
